+++ Nexus/DataStructures/segtree.cpp
@@ -5,7 +5,7 @@
 // INTERFACE CONTRACT
 // ===================
 // template<typename T>
-// concept SegmentTreeConfig = requires(typename T::Node n, typename T::Lazy l, int len) {
+// concept SegmentTreeConfig = requires(typename T::Node n, typename T::Lazy l, ll len) {
 //     typename T::Node;
 //     typename T::Lazy;
 //
@@ -31,29 +31,29 @@ class SegmentTree {
 public:
     explicit SegmentTree(const vector<Node>& input) {
         n = input.size();
-        size = bit_ceil(static_cast<unsigned int>(n));
+        size = bit_ceil(static_cast<ull>(n));
 
         data.assign(2 * size, Node::identity());
         lazy.assign(size, Lazy::identity());
         segLen.assign(2 * size, 0);
 
-        for(int i = 0; i < n; i++) {
+        for(ll i = 0; i < n; i++) {
             data[size + i] = input[i];
             segLen[size + i] = 1;
         }
 
-        for(int i = size - 1; i >= 1; i--) {
+        for(ll i = size - 1; i >= 1; i--) {
             data[i] = Config::merge(data[2 * i], data[2 * i + 1]);
             segLen[i] = segLen[2 * i] + segLen[2 * i + 1];
         }
     }
 
-    void update(int left, int right, Lazy func) {
+    void update(ll left, ll right, Lazy func) {
         left += size;
         right += size;
 
-        int leftCopy = left;
-        int rightCopy = right;
+        ll leftCopy = left;
+        ll rightCopy = right;
 
         pushPath(leftCopy);
         pushPath(rightCopy);
@@ -70,7 +70,7 @@ public:
         buildPath(rightCopy);
     }
 
-    Node query(int left, int right) {
+    Node query(ll left, ll right) {
         left += size;
         right += size;
 
@@ -92,19 +92,19 @@ public:
     }
 
 private:
-    int n;
-    int size;
+    ll n;
+    ll size;
     vector<Node> data;
     vector<Lazy> lazy;
-    vector<int> segLen;
+    vector<ll> segLen;
 
-    inline void applyNode(int pos, const Lazy& func) {
+    inline void applyNode(ll pos, const Lazy& func) {
         data[pos] = Config::apply(func, data[pos], segLen[pos]);
 
         if(pos < size) lazy[pos] = Config::compose(func, lazy[pos]);
     }
 
-    inline void push(int pos) {
+    inline void push(ll pos) {
         if(!(lazy[pos] == Lazy::Identity())) {
             applyNode((pos << 1), lazy[pos]);
             applyNode((pos << 1) | 1, lazy[pos]);
@@ -113,15 +113,15 @@ private:
         }
     }
 
-    inline void pushPath(int pos) {
-        int height = countr_zero(static_cast<unsigned int>(size));
+    inline void pushPath(ll pos) {
+        ll height = countr_zero(static_cast<ull>(size));
 
-        for(int h = height; h >= 1; h--) {
+        for(ll h = height; h >= 1; h--) {
             push(pos >> h);
         }
     }
 
-    inline void buildPath(int pos) {
+    inline void buildPath(ll pos) {
         while(pos > 1) {
             pos >>= 1;
 
@@ -141,7 +141,7 @@ private:
 // 1. Range Mul-Add + Range Sum
 struct RMARS {
     struct Node {
-        int val;
+        ll val;
 
         static Node identity() {
             return Node{ 0 };
@@ -151,8 +151,8 @@ struct RMARS {
     };
 
     struct Lazy {
-        int mul;
-        int add;
+        ll mul;
+        ll add;
 
         static Lazy identity() {
             return Lazy{ 1, 0 };
@@ -165,7 +165,7 @@ struct RMARS {
         return Node{ a.val + b.val };
     }
 
-    static Node apply(const Lazy& f, const Node& node, int len) {
+    static Node apply(const Lazy& f, const Node& node, ll len) {
         return Node{ node.val * f.mul + f.add * len };
     }
 
@@ -177,7 +177,7 @@ struct RMARS {
 // 2. Range Add + Range Sum
 struct RARS {
     struct Node {
-        int sum;
+        ll sum;
 
         static Node identity() {
             return Node{ 0 };
@@ -187,7 +187,7 @@ struct RARS {
     };
 
     struct Lazy {
-        int add;
+        ll add;
 
         static Lazy identity() {
             return Lazy{ 0 };
@@ -200,7 +200,7 @@ struct RARS {
         return Node{ a.sum + b.sum };
     }
 
-    static Node apply(const Lazy& f, const Node& node, int len) {
+    static Node apply(const Lazy& f, const Node& node, ll len) {
         return Node{ node.sum + f.add * len };
     }
 
@@ -212,17 +212,17 @@ struct RARS {
 // 3. Range Add + Range Max
 struct RARM {
     struct Node {
-        int mx;
+        ll mx;
 
         static Node identity() {
-            return Node{ -INF<int> };
+            return Node{ -INF<ll> };
         }
 
         bool operator==(const Node&) const = default;
     };
 
     struct Lazy {
-        int add;
+        ll add;
 
         static Lazy identity() {
             return Lazy{ 0 };
@@ -235,8 +235,8 @@ struct RARM {
         return Node{ max(a.mx, b.mx) };
     }
 
-    static Node apply(const Lazy& f, const Node& node, int len) {
-        if(node.mx == INF<int>) return node;
+    static Node apply(const Lazy& f, const Node& node, ll len) {
+        if(node.mx == INF<ll>) return node;
 
         return Node{ node.mx + f.add };
     }
@@ -249,8 +249,8 @@ struct RARM {
 // 4. Range Add + Range Sum-SumOfSquares
 struct RASS {
     struct Node {
-        int sum;
-        int sq;
+        ll sum;
+        ll sq;
 
         static Node identity() {
             return Node{ 0, 0 };
@@ -260,7 +260,7 @@ struct RASS {
     };
 
     struct Lazy {
-        int add;
+        ll add;
 
         static Lazy identity() {
             return Lazy{0};
@@ -273,9 +273,9 @@ struct RASS {
         return Node{ a.sum + b.sum, a.sq + b.sq };
     }
 
-    static Node apply(const Lazy& f, const Node& node, int len) {
-        int new_sum = node.sum + f.add * len;
-        int new_sq = node.sq + 2LL * f.add * node.sum + f.add * f.add * len;
+    static Node apply(const Lazy& f, const Node& node, ll len) {
+        ll new_sum = node.sum + f.add * len;
+        ll new_sq = node.sq + 2LL * f.add * node.sum + f.add * f.add * len;
 
         return Node{ new_sum, new_sq };
     }
@@ -288,18 +288,18 @@ struct RASS {
 // 5. No Lazy + DP Transition
 struct NLDP {
     struct Node {
-        array<array<int, 2>, 2> m;
+        array<array<ll, 2>, 2> m;
 
         static Node identity() {
             return Node{{{
-                { 0, INF<int> },
-                { INF<int>, 0 }
+                { 0, INF<ll> },
+                { INF<ll>, 0 }
             }}};
         }
 
-        static Node fromVal(int a) {
+        static Node fromVal(ll a) {
             return Node{{{
-                { INF<int>, a },
+                { INF<ll>, a },
                 { 0, a }
             }}};
         }
@@ -318,11 +318,11 @@ struct NLDP {
     static Node merge(const Node& a, const Node& b) {
         Node res;
 
-        for(int i = 0; i < 2; i++) {
-            for(int j = 0; j < 2; j++) {
-                res.m[i][j] = INF<int>;
+        for(ll i = 0; i < 2; i++) {
+            for(ll j = 0; j < 2; j++) {
+                res.m[i][j] = INF<ll>;
 
-                for(int k = 0; k < 2; k++) {
+                for(ll k = 0; k < 2; k++) {
                     chkmin(res.m[i][j], a.m[i][k] + b.m[k][j]);
                 }
             }
@@ -331,7 +331,7 @@ struct NLDP {
         return res;
     }
 
-    static Node apply(const Lazy& f, const Node& node, int len) {
+    static Node apply(const Lazy& f, const Node& node, ll len) {
         return node;
     }
